+++ examples/simple/customauth/__init__.py
@@ -1,0 +1,1 @@
+from customauth.models import MyUser