+++ src/fobi/tests/base.py
@@ -12,7 +12,6 @@ from time import sleep
 from copy import copy
 
 from django.test import TestCase, LiveServerTestCase, Client, RequestFactory
-#from django.contrib.auth.models import User
 from django.core.management import call_command
 from django.contrib.staticfiles.management.commands import collectstatic
 from django.conf import settings
@@ -27,46 +26,48 @@ from fobi.base import (
 from fobi.models import FormEntry, FormElementEntry, FormHandlerEntry
 from fobi.dynamic import assemble_form_class
 
-from fobi.contrib.plugins.form_elements.content.content_text.fobi_form_elements import ContentTextPlugin
-from fobi.contrib.plugins.form_elements.content.content_image.fobi_form_elements import ContentImagePlugin
+from fobi.contrib.plugins.form_elements.content.content_text.fobi_form_elements \
+    import ContentTextPlugin
+from fobi.contrib.plugins.form_elements.content.content_image.fobi_form_elements \
+    import ContentImagePlugin
 
-from fobi.contrib.plugins.form_elements.fields.boolean.fobi_form_elements import BooleanSelectPlugin
-from fobi.contrib.plugins.form_elements.fields.email.fobi_form_elements import EmailPlugin
-from fobi.contrib.plugins.form_elements.fields.hidden.fobi_form_elements import HiddenInputPlugin
-from fobi.contrib.plugins.form_elements.fields.integer.fobi_form_elements import IntegerInputPlugin
-from fobi.contrib.plugins.form_elements.fields.text.fobi_form_elements import TextInputPlugin
-from fobi.contrib.plugins.form_elements.fields.textarea.fobi_form_elements import TextareaPlugin
+from fobi.contrib.plugins.form_elements.fields.boolean.fobi_form_elements \
+    import BooleanSelectPlugin
+from fobi.contrib.plugins.form_elements.fields.email.fobi_form_elements \
+    import EmailPlugin
+from fobi.contrib.plugins.form_elements.fields.hidden.fobi_form_elements \
+    import HiddenInputPlugin
+from fobi.contrib.plugins.form_elements.fields.integer.fobi_form_elements \
+    import IntegerInputPlugin
+from fobi.contrib.plugins.form_elements.fields.text.fobi_form_elements \
+    import TextInputPlugin
+from fobi.contrib.plugins.form_elements.fields.textarea.fobi_form_elements \
+    import TextareaPlugin
 
-from fobi.contrib.plugins.form_handlers.db_store.fobi_form_handlers import DBStoreHandlerPlugin
-from fobi.contrib.plugins.form_handlers.mail.fobi_form_handlers import MailHandlerPlugin
+from fobi.contrib.plugins.form_handlers.db_store.fobi_form_handlers \
+    import DBStoreHandlerPlugin
+from fobi.contrib.plugins.form_handlers.mail.fobi_form_handlers \
+    import MailHandlerPlugin
 
 # ****************************************************************************
 # **************** Safe User import for Django > 1.5, < 1.8 ******************
 # ****************************************************************************
-try:
-    # Django 1.7 check
-    from django.apps import AppConfig
-    from django.conf import settings
-    User = settings.AUTH_USER_MODEL
-except ImportError:
-    # Django 1.6 check
-    try:
-        from django.contrib.auth import get_user_model
-    # Fall back to Django 1.5
-    except ImportError:
-        from django.contrib.auth.models import User
-    else:
-        User = get_user_model()
+AUTH_USER_MODEL = settings.AUTH_USER_MODEL
 
-    # Sanity checks
-    user = User()
+# Note, that this may cause circular imports - thus the ``get_user_model``
+# should be moved elsewhere (be used on the function/method level). For
+# now leave commented and solve in future. Possible use the DjangoCMS solution
+# https://github.com/divio/django-cms/blob/develop/cms/models/permissionmodels.py#L18
 
-    if not hasattr(user, 'username'):
-        from fobi.exceptions import ImproperlyConfigured
-        raise ImproperlyConfigured("Your custom user model ({0}.{1}) doesn't "
-                                   "have ``username`` property, while "
-                                   "``django-fobi`` relies on its' presence"
-                                   ".".format(user._meta.app_label, user._meta.object_name))
+# Sanity checks.
+#user = User()
+#
+#if not hasattr(user, 'username'):
+#    from fobi.exceptions import ImproperlyConfigured
+#    raise ImproperlyConfigured("Your custom user model ({0}.{1}) doesn't "
+#                               "have ``username`` property, while "
+#                               "``django-fobi`` relies on its' presence"
+#                               ".".format(user._meta.app_label, user._meta.object_name))
 
 # ****************************************************************************
 # ****************************************************************************